app/rds_init_lambda/requirements.txt
mysql-connector-python==8.0.33
aws-lambda-powertools==1.28.0
python-dotenv==0.19.0